Add commonServerMain scope.
